+++ src/api/integrations/event/event.router.ts
@@ -4,6 +4,8 @@ import { WebhookRouter } from '@api/integrations/event/webhook/webhook.router';
 import { WebsocketRouter } from '@api/integrations/event/websocket/websocket.router';
 import { Router } from 'express';
 
+import { PusherRouter } from './pusher/pusher.router';
+
 export class EventRouter {
   public readonly router: Router;
 
@@ -13,6 +15,7 @@ export class EventRouter {
     this.router.use('/webhook', new WebhookRouter(configService, ...guards).router);
     this.router.use('/websocket', new WebsocketRouter(...guards).router);
     this.router.use('/rabbitmq', new RabbitmqRouter(...guards).router);
+    this.router.use('/pusher', new PusherRouter(...guards).router);
     this.router.use('/sqs', new SqsRouter(...guards).router);
   }
 }